# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CatalogItemOut": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "availability": {
            "title": "Availability",
            "type": "string"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "categoryAr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Categoryar"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "descriptionAr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descriptionar"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "nameAr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Namear"
          },
          "price": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "priceLastSyncedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pricelastsyncedat"
          },
          "sku": {
            "title": "Sku",
            "type": "string"
          },
          "stale": {
            "title": "Stale",
            "type": "boolean"
          },
          "unitLabel": {
            "title": "Unitlabel",
            "type": "string"
          }
        },
        "required": [
          "id",
          "sku",
          "name",
          "description",
          "category",
          "unitLabel",
          "active",
          "availability",
          "price",
          "currency",
          "priceLastSyncedAt",
          "stale"
        ],
        "title": "CatalogItemOut",
        "type": "object"
      },
      "CategoryOut": {
        "properties": {
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "value"
        ],
        "title": "CategoryOut",
        "type": "object"
      },
      "CreateOrderRequest": {
        "properties": {
          "customerPhone": {
            "anyOf": [
              {
                "maxLength": 24,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customerphone"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderItemRequest"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Items",
            "type": "array"
          },
          "note": {
            "default": "",
            "maxLength": 500,
            "title": "Note",
            "type": "string"
          },
          "storeId": {
            "maxLength": 36,
            "minLength": 1,
            "title": "Storeid",
            "type": "string"
          }
        },
        "required": [
          "storeId",
          "items"
        ],
        "title": "CreateOrderRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LoginRequest": {
        "properties": {
          "email": {
            "maxLength": 255,
            "minLength": 3,
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 200,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "LoginResponse": {
        "properties": {
          "accessToken": {
            "title": "Accesstoken",
            "type": "string"
          },
          "expiresIn": {
            "title": "Expiresin",
            "type": "integer"
          },
          "tokenType": {
            "title": "Tokentype",
            "type": "string"
          }
        },
        "required": [
          "accessToken",
          "tokenType",
          "expiresIn"
        ],
        "title": "LoginResponse",
        "type": "object"
      },
...
      "OrderItemOut": {
        "properties": {
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "lineTotal": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Linetotal",
            "type": "string"
          },
          "productId": {
            "title": "Productid",
            "type": "string"
          },
          "productName": {
            "title": "Productname",
            "type": "string"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "unitPrice": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Unitprice",
            "type": "string"
          }
        },
        "required": [
          "productId",
          "productName",
          "quantity",
          "unitPrice",
          "currency",
          "lineTotal"
        ],
        "title": "OrderItemOut",
        "type": "object"
      },
      "OrderItemRequest": {
        "properties": {
          "clientUnitPrice": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clientunitprice"
          },
          "productId": {
            "title": "Productid",
            "type": "string"
          },
          "quantity": {
            "exclusiveMinimum": 0.0,
            "maximum": 200.0,
            "title": "Quantity",
            "type": "integer"
          }
        },
        "required": [
          "productId",
          "quantity"
        ],
        "title": "OrderItemRequest",
        "type": "object"
      },
      "OrderOut": {
        "properties": {
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "customerPhone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customerphone"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderItemOut"
            },
            "title": "Items",
            "type": "array"
          },
          "publicOrderCode": {
            "title": "Publicordercode",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "storeId": {
            "title": "Storeid",
            "type": "string"
          },
          "submittedAt": {
            "format": "date-time",
            "title": "Submittedat",
            "type": "string"
          },
          "totalAmount": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Totalamount",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "publicOrderCode",
          "storeId",
          "state",
          "totalAmount",
          "currency",
          "version",
          "submittedAt",
          "items"
        ],
        "title": "OrderOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Offline-first Click & Collect portfolio API.",
    "title": "Jahhezly API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/auth/login": {
      "post": {
        "operationId": "login_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/v1/orders": {
      "post": {
        "operationId": "create_order_v1_orders_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "X-Guest-Session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Guest-Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Order",
        "tags": [
          "customer"
        ]
      }
    },
    "/v1/orders/{order_id}": {
      "get": {
        "operationId": "get_order_v1_orders__order_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "title": "Order Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Guest-Session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Guest-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Order",
        "tags": [
          "customer"
        ]
      }
    },
    "/v1/orders/{order_id}/cancel": {
      "post": {
        "operationId": "cancel_order_v1_orders__order_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "title": "Order Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "title": "If-Match",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Guest-Session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Guest-Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Order",
        "tags": [
          "customer"
        ]
      }
    },
...
    "/v1/stores/{store_id}/catalog": {
      "get": {
        "operationId": "catalog_v1_stores__store_id__catalog_get",
        "parameters": [
          {
            "in": "path",
            "name": "store_id",
            "required": true,
            "schema": {
              "title": "Store Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CatalogItemOut"
                  },
                  "title": "Response Catalog V1 Stores  Store Id  Catalog Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Catalog",
        "tags": [
          "customer"
        ]
      }
    },
    "/v1/stores/{store_id}/categories": {
      "get": {
        "operationId": "categories_v1_stores__store_id__categories_get",
        "parameters": [
          {
            "in": "path",
            "name": "store_id",
            "required": true,
            "schema": {
              "title": "Store Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CategoryOut"
                  },
                  "title": "Response Categories V1 Stores  Store Id  Categories Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Categories",
        "tags": [
          "customer"
        ]
      }
    }
  }
}